requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
gevent==24.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
setuptools==69.5.1
urllib3==2.2.1
Werkzeug==3.0.2
zope.event==5.0
zope.interface==6.3
